ref: fix type of Group.priority (#91874)
it is integer in production, this table is far too large to rewrite so
adjust the model definition to match

<!-- Describe your PR here. -->

+++ src/sentry/models/group.py
@@ -600,7 +600,7 @@ class Group(Model):
     type = BoundedPositiveIntegerField(
         default=DEFAULT_TYPE_ID, db_default=DEFAULT_TYPE_ID, db_index=True
     )
-    priority = models.PositiveSmallIntegerField(null=True)
+    priority = models.PositiveIntegerField(null=True)
     priority_locked_at = models.DateTimeField(null=True)
     seer_fixability_score = models.FloatField(null=True)
     seer_autofix_last_triggered = models.DateTimeField(null=True)
